+++ test/e2e/elements.js
@@ -46,6 +46,10 @@ module.exports = {
     },
   },
   screens: {
+    eventsEmitter: {
+      button: By.className('btn btn-default'),
+      event: By.className('Toastify__toast-body'),
+    },
     signMessage: {
       buttons: {
         sign: By.css('#app-content > div > div.app-primary.from-right > div > div > div.flex-row.flex-space-around > button:nth-child(2)'),
@@ -161,6 +165,10 @@ module.exports = {
       button: {
         submit: By.css('#pending-tx-form > div.flex-row.flex-space-around.conf-buttons > input'),
       },
+      fields: {
+        gasLimit: By.css('#pending-tx-form > div:nth-child(1) > div.table-box > div:nth-child(3) > div.cell.value > div > div > input'),
+        gasPrice: By.css('#pending-tx-form > div:nth-child(1) > div.table-box > div:nth-child(4) > div.cell.value > div > div > input'),
+      },
     },
     sendTransaction: {
       title: By.css('#app-content > div > div.app-primary.from-right > div > h3:nth-child(2)'),
@@ -233,7 +241,8 @@ module.exports = {
         menu: By.className('wallet-view__tab-history'),
         tokens: By.className('activeForm right'),
       },
-      balance: By.css('#app-content > div > div.app-primary.from-right > div > div > div.flex-row > div.ether-balance.ether-balance-amount > div > div > div:nth-child(1) > div:nth-child(1)'),
+     // balance: By.css('#app-content > div > div.app-primary.from-right > div > div > div.flex-row > div.ether-balance.ether-balance-amount > div > div > div:nth-child(1) > div:nth-child(1)'),
+     balance: By.xpath('//*[@id="app-content"]/div/div[2]/div/div/div[2]/div[1]/div/div/div[1]/div[1]'),
       address: By.css('#app-content > div > div.app-primary.from-left > div > div > div:nth-child(1) > flex-column > div.flex-row > div'),
       tokens: {
         menu: By.id('wallet-view__tab-tokens'),